+++ pyvips/base.py
@@ -120,15 +120,18 @@ def values_for_enum(gtype):
     g_type_class = gobject_lib.g_type_class_ref(gtype)
     g_enum_class = ffi.cast('GEnumClass *', g_type_class)
 
-    values = []
-
     # -1 since we always have a "last" member.
-    for i in range(0, g_enum_class.n_values - 1):
-        value = _to_string(g_enum_class.values[i].value_nick)
-        values.append(value)
+    return [_to_string(g_enum_class.values[i].value_nick)
+            for i in range(g_enum_class.n_values - 1)]
 
-    return values
+def values_for_flag(gtype):
+    """Get all values for a flag (gtype)."""
 
+    g_type_class = gobject_lib.g_type_class_ref(gtype)
+    g_flags_class = ffi.cast('GFlagsClass *', g_type_class)
+
+    return [_to_string(g_flags_class.values[i].value_nick)
+            for i in range(g_flags_class.n_values)]
 
 __all__ = [
     'leak_set',
@@ -142,5 +145,6 @@ __all__ = [
     'type_name',
     'type_map',
     'type_from_name',
-    'values_for_enum'
+    'values_for_enum',
+    'values_for_flag'
 ]